Middleware

+++ routes/web.php
@@ -7,13 +7,11 @@ use App\Http\Controllers\StudentController;
 use App\Http\Controllers\TeacherController;
 use App\Http\Controllers\ExtracurricularController;
 
-Route::get('/', function () {
-    return view('welcome');
-});
+
 
 Route::get('/home', function () {
     return view('home');
-});
+})->middleware('auth');
 
 Route::get('/login', ([AuthController::class, 'login']))->name('login')->middleware('guest');
 Route::post('/login', ([AuthController::class, 'authenticating']))->middleware('guest');
@@ -21,15 +19,15 @@ Route::post('/login', ([AuthController::class, 'authenticating']))->middleware('
 Route::get('/logout', ([AuthController::class, 'logout']))->middleware('auth');
 
 Route::get('/student', ([StudentController::class, 'index']))->middleware('auth');
-Route::get('/student-detail/{id}', ([StudentController::class, 'show']))->middleware('auth');
-Route::get('/student-add', ([StudentController::class, 'create']))->middleware('auth');
-Route::post('/student', ([StudentController::class, 'store']))->middleware('auth');
-Route::get('/student-edit/{id}', ([StudentController::class, 'edit']))->middleware('auth');
-Route::put('/student/{id}', ([StudentController::class, 'update']))->middleware('auth');
-Route::get('/student-delete/{id}', ([StudentController::class, 'delete']))->middleware('auth');
-Route::delete('/student-destroy/{id}', ([StudentController::class, 'destroy']))->middleware('auth');
-Route::get('/student-deleted', ([StudentController::class, 'deletedStudent']))->middleware('auth');
-Route::get('/student/{id}/restore', ([StudentController::class, 'restore']))->middleware('auth');
+Route::get('/student-detail/{id}', ([StudentController::class, 'show']))->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/student-add', ([StudentController::class, 'create']))->middleware(['auth', 'must-admin-or-teacher']);
+Route::post('/student', ([StudentController::class, 'store']))->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/student-edit/{id}', ([StudentController::class, 'edit']))->middleware(['auth', 'must-admin-or-teacher']);
+Route::put('/student/{id}', ([StudentController::class, 'update']))->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/student-delete/{id}', ([StudentController::class, 'delete']))->middleware(['auth', 'must-admin']);
+Route::delete('/student-destroy/{id}', ([StudentController::class, 'destroy']))->middleware(['auth', 'must-admin']);
+Route::get('/student-deleted', ([StudentController::class, 'deletedStudent']))->middleware(['auth', 'must-admin']);
+Route::get('/student/{id}/restore', ([StudentController::class, 'restore']))->middleware(['auth', 'must-admin']);
 
 Route::get('/classroom', ([ClassController::class, 'index']))->middleware('auth');
 Route::get('/classroom-detail/{id}', ([ClassController::class, 'show']))->middleware('auth');
